Almost done fixing.

+++ gettsim/demographic_vars.py
@@ -3,7 +3,7 @@ def _anz_kinder_in_tu(tu_id, kind):
 
 
 def _anz_erwachsene_tu(tu_id, kind):
-    return (~kind).groupby(tu_id).sum()
+    return (~kind).astype(int).groupby(tu_id).sum()
 
 
 def gemeinsam_veranlagt(tu_id, _anz_erwachsene_tu):
